DP-2402 - Refactor preview rendering to test it during feature enable

+++ app/BUILD/src/DeskPRO/Bundle/AppBundle/Features/EmailTemplatesFeature.php
@@ -9,7 +9,10 @@ use Application\DeskPRO\Templating\Templates\EmailTemplateCode;
 use Application\DeskPRO\Templating\Templates\TemplateCode;
 use Application\DeskPRO\Templating\Templates\TemplateCustom;
 use DeskPRO\Bundle\AppBundle\Templating\EmailTemplatesDesc;
+use DeskPRO\Bundle\SendmailBundle\Factory\AgentViewModelFactory;
+use DeskPRO\Bundle\SendmailBundle\Factory\UserViewModelFactory;
 use DeskPRO\Bundle\SendmailBundle\Templating\Templates\TemplateSet;
+use DeskPRO\Bundle\SendmailBundle\View\Model\EmailBaseType;
 use Doctrine\ORM\EntityManager;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -165,6 +168,12 @@ HTML;
         $templatesDesc = new EmailTemplatesDesc();
         $manifest      = $templatesDesc->getManifest();
 
+        $renderer = $container->get('email.email_renderer');
+
+        $dataFactory = $container->get('email.preview_fake_data_factory');
+
+        $language = $container->get('language_manager')->getLanguageStack()->getDefaultLanguage();
+
         foreach ($templates as $previousTemplate) {
             /** @var Template $previousTemplate */
             if (strpos($previousTemplate->getName(), 'DeskPRO:emails_custom') === 0) {
@@ -204,8 +213,57 @@ HTML;
             if (!$body) {
                 continue;
             }
-            $this->migratedCustomTemplates[$previousTemplate->getName()] = $newTemplateName;
             $templateCode->setBody($body);
+
+            $subject = $this->convertTemplateCode($templateCode->getSubject(), true);
+            if (!$subject) {
+                continue;
+            }
+            $templateCode->setSubject($subject);
+
+            // Loader issue cause this check to fail for all templates, I keep it commented until I sort it out.
+//            try {
+//                if (strpos($newTemplateName, 'SendmailBundle:emails_custom:') === 0) {
+//                    $factory   = $container->get('email.custom_viewmodel_factory');
+//                    $arguments = $dataFactory->getArguments($factory, 'createCustomTemplateModel', null);
+//                    $model = call_user_func_array([$factory, 'createCustomTemplateModel'], $arguments);
+//                } else {
+//                    /** @var EmailBaseType $model */
+//                    $templatesDesc = new EmailTemplatesDesc();
+//                    $manifest      = $templatesDesc->getManifest();
+//                    $viewModel     = false;
+//                    foreach ($manifest as $t) {
+//                        if (isset($t['newTemplate']) && $t['newTemplate'] === $newTemplateName) {
+//                            if ($t['viewModel']) {
+//                                $viewModel = $t['viewModel'];
+//                            }
+//                            break;
+//                        }
+//                    }
+
+//                    /** @var AgentViewModelFactory|UserViewModelFactory $factory */
+//                    $factory = strpos($newTemplateName, 'SendmailBundle:emails_agent:') === 0
+//                        ? $container->get('email.agent_viewmodel_factory')
+//                        : $container->get('email.user_viewmodel_factory');
+//                    $action  = 'create'.$viewModel.'Model';
+
+//                    $arguments = $dataFactory->getArguments($factory, $action, null);
+
+//                    if (!is_callable([$factory, $action])) {
+//                        return false;
+//                    }
+
+//                    $model = call_user_func_array([$factory, $action], $arguments);
+//                }
+
+//                $tplName = uniqid('SendmailBundle:emails_', true);
+
+//                $renderer->renderPreview($templateCode->getCode(), $tplName, $model, $language, $templates);
+//            } catch (\Throwable $e) {
+//                continue;
+//            }
+
+            $this->migratedCustomTemplates[$previousTemplate->getName()] = $newTemplateName;
             $set->saveTemplate($template);
             $this->saveLegacyTemplate($em, $previousTemplate);
             $em->remove($previousTemplate);
@@ -226,7 +284,13 @@ $code
 CODE;
     }
 
-    private function convertTemplateCode($code)
+    /**
+     * @param string $code
+     * @param bool   $subject
+     *
+     * @return bool|mixed|null|string|string[]
+     */
+    private function convertTemplateCode($code, $subject = false)
     {
         $code = $this->uniformiseVariableSyntax($code);
         $code = str_replace(
@@ -237,6 +301,7 @@ CODE;
         if (preg_match('/<dp:/', $code)) {
             return false;
         }
+        $code      = preg_replace('/{{\s*(phrase\s*\([^{]+\))\|\s*raw\s*}}/', '{{ $1 }}', $code);
         $whiteList = $this->getVariableWhiteList();
         if (preg_match_all('/{{[^}]+}}/', $code, $matches)) {
             foreach ($matches[0] as $match) {
@@ -251,6 +316,11 @@ CODE;
                 }
             }
         }
+
+        if ($subject) {
+            return $code;
+        }
+
         $code = preg_replace('/.+/', '    $0', $code);
 
         return <<<CODE
